+++ dev/sparktestsupport/modules.py
@@ -862,7 +862,6 @@ pyspark_pandas_slow = Module(
         "pyspark.pandas.tests.groupby.test_stat_prod",
         "pyspark.pandas.tests.groupby.test_value_counts",
         "pyspark.pandas.tests.test_indexing",
-        "pyspark.pandas.tests.test_ops_on_diff_frames",
         "pyspark.pandas.tests.diff_frames_ops.test_align",
         "pyspark.pandas.tests.diff_frames_ops.test_arithmetic",
         "pyspark.pandas.tests.diff_frames_ops.test_arithmetic_ext",
@@ -872,6 +871,7 @@ pyspark_pandas_slow = Module(
         "pyspark.pandas.tests.diff_frames_ops.test_arithmetic_chain_ext_float",
         "pyspark.pandas.tests.diff_frames_ops.test_assign_frame",
         "pyspark.pandas.tests.diff_frames_ops.test_assign_series",
+        "pyspark.pandas.tests.diff_frames_ops.test_basic",
         "pyspark.pandas.tests.diff_frames_ops.test_bitwise",
         "pyspark.pandas.tests.diff_frames_ops.test_combine_first",
         "pyspark.pandas.tests.diff_frames_ops.test_compare_series",
@@ -1235,7 +1235,6 @@ pyspark_pandas_connect_part3 = Module(
         "pyspark.pandas.tests.connect.indexes.test_parity_datetime_map",
         "pyspark.pandas.tests.connect.indexes.test_parity_datetime_property",
         "pyspark.pandas.tests.connect.indexes.test_parity_datetime_round",
-        "pyspark.pandas.tests.connect.test_parity_ops_on_diff_frames",
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_arithmetic",
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_arithmetic_ext",
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_arithmetic_ext_float",
@@ -1244,6 +1243,7 @@ pyspark_pandas_connect_part3 = Module(
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_arithmetic_chain_ext_float",
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_assign_frame",
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_assign_series",
+        "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_basic",
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_bitwise",
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_combine_first",
         "pyspark.pandas.tests.connect.diff_frames_ops.test_parity_compare_series",
