+++ python/pyspark/pandas/tests/connect/diff_frames_ops/test_parity_basic.py
@@ -16,20 +16,21 @@
 #
 import unittest
 
-from pyspark.pandas.tests.test_ops_on_diff_frames import OpsOnDiffFramesEnabledTestsMixin
-
 from pyspark.testing.connectutils import ReusedConnectTestCase
 from pyspark.testing.pandasutils import PandasOnSparkTestUtils
+from pyspark.pandas.tests.diff_frames_ops.test_basic import BasicMixin
 
 
-class OpsOnDiffFramesEnabledParityTests(
-    OpsOnDiffFramesEnabledTestsMixin, PandasOnSparkTestUtils, ReusedConnectTestCase
+class BasicParityTests(
+    BasicMixin,
+    PandasOnSparkTestUtils,
+    ReusedConnectTestCase,
 ):
     pass
 
 
 if __name__ == "__main__":
-    from pyspark.pandas.tests.connect.test_parity_ops_on_diff_frames import *  # noqa: F401
+    from pyspark.pandas.tests.connect.diff_frames_ops.test_parity_basic import *  # noqa: F401
 
     try:
         import xmlrunner  # type: ignore[import]
